fix: use discriminated union for variants of CollectionCard
