+++ src/common/index.ts
@@ -1,8 +1,8 @@
-export { default as ArticleCardProps } from "./ArticleCard"
 export { default as BreadcrumbProps } from "./Breadcrumb"
 export { default as ButtonProps } from "./Button"
 export { default as CalloutProps } from "./Callout"
 export { default as CardsProps } from "./Cards"
+export { default as CollectionCardProps } from "./CollectionCard"
 export { default as ContentProps } from "./Content"
 export { default as ContentPageHeaderProps } from "./ContentPageHeader"
 export { default as FooterProps } from "./Footer"
