--- src/templates/next/components/ArticleCard/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./ArticleCard"
