+++ src/templates/next/components/CollectionCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./CollectionCard"
