+++ src/templates/next/components/index.ts
@@ -1,7 +1,7 @@
-export { default as ArticleCard } from "./ArticleCard"
 export { default as Breadcrumb } from "./Breadcrumb"
 export { default as Button } from "./Button"
 export { default as Callout } from "./Callout"
+export { default as CollectionCard } from "./CollectionCard"
 export { default as ContentPageHeader } from "./ContentPageHeader"
 export { default as Footer } from "./Footer"
 export { default as Heading } from "./Heading"
